Move Aurora entity to separate file
